+++ gDRplots/NEWS.md
@@ -1,5 +1,11 @@
-## gDRplots 0.0.6 - 2024-06-19
+## gDRplots 0.0.8 - 2024-07-01
 * add function for plot size estimation
+
+## gDRplots 0.0.7 - 2024-06-21
+* add plot to quality check of dose response
+
+## gDRplots 0.0.6 - 2024-06-19
+* update plot function for `processingReport` package
 
 ## gDRplots 0.0.5 - 2024-06-14
 * move plot function from `metricClustering`, `responseGrid` and `metricDistribution` packages